move redirect to util to keep express-routes isolated
